feat: implement custom JWT authentication converter and user auditor aware functionality

+++ back/src/main/java/com/openclassrooms/mddapi/features/comment/CommentResponseDto.java
@@ -1,7 +1,7 @@
 package com.openclassrooms.mddapi.features.comment;
 
 public interface CommentResponseDto {
-    Long getId();
+    Integer getId();
 
     String getMessage();
